requirements.txt
# Core dependencies
aiohttp==3.9.3
beautifulsoup4==4.12.3
playwright==1.42.0
pydantic==2.6.3
pydantic[email]==2.6.3
typer[all]==0.9.0
rich==13.7.1
python-dotenv==1.0.1

# Database
# SQLite is built into Python, no additional package needed

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development
black==24.2.0
flake8==7.0.0
mypy==1.9.0
types-beautifulsoup4==4.12.0.20240229
types-aiofiles==23.2.0.20240311

# Optional - for better async performance
uvloop==0.19.0 ; platform_system != "Windows"
